Shuffle some code around to allow CLI commands to run without dependencies where possible Add CLI command for XA.MXA export Fix VAB CLI command not actually running the selected command
